Refactor use-pagination hook to improve code readability and maintainability
